+++ src/components/Container/StudyroomWaitContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import axios from "axios";
 import { useNavigate, useParams } from "react-router-dom";
-import LargeUserDisplay from "../LargeUserDisplay";
+import LargeUserDisplay from "../rooms/LargeUserDisplay";
 
 type StudyroomWaitContainerProps = {};
