+++ src/components/rooms/CategorySelector.tsx
