+++ src/components/rooms/LargeUserDisplay.tsx
