+++ src/components/rooms/SmallUserDisplay.tsx
