+++ src/components/rooms/Studyroom.tsx
