+++ src/components/rooms/StudyroomFooter.tsx
