+++ src/components/rooms/StudyroomHeader.tsx
