+++ src/components/rooms/StudyroomTN.tsx
