+++ src/components/users/RankingSlider.tsx
