+++ src/pages/StudyRecruitListPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Header from '../components/Header';
+import Header from '../components/common/Header';
 import RecruitStudyListContainer from '../components/Container/RecruitStudyListContainer'
-import Footer from "../components/Footer";
+import Footer from "../components/common/Footer";
 
 type RecruitStudyListPageProps = {};
